// vietnam/assets/Tienlen/script/protocol/tournament.d.ts
export namespace tournament {
// ...
    interface IItemInfo {

        /** ItemInfo ItemId */
        ItemId?: (number | null);

        /** ItemInfo ItemNum */
        ItemNum?: (number | null);

        /** ItemInfo Name */
        Name?: (string | null);
    }
// ...
    interface IMatchInfoAward {

        /** MatchInfoAward ItemInfo */
        ItemInfo?: (tournament.IItemInfo[] | null);

        /** MatchInfoAward Coin */
        Coin?: (number | Long | null);

        /** MatchInfoAward Diamond */
        Diamond?: (number | Long | null);

        /** MatchInfoAward UpLimit */
        UpLimit?: (number | null);

        /** MatchInfoAward DownLimit */
        DownLimit?: (number | null);
    }
// ...
}